added tailwind css and authentication functionality and views

+++ webpack.mix.js
@@ -1,4 +1,5 @@
 const mix = require('laravel-mix');
+
 require('laravel-mix-tailwind');
 require('laravel-mix-purgecss');
 
@@ -14,6 +15,11 @@ require('laravel-mix-purgecss');
  */
 
 mix.js('resources/js/app.js', 'public/js')
-    .sass('resources/sass/app.scss', 'public/css')
-    .tailwind()
-    .purgeCss();
+   .postCss('resources/css/app.css', 'public/css')
+   .tailwind('./tailwind.config.js');
+
+if (mix.inProduction()) {
+  mix
+   .version()
+   .purgeCss();
+}
